+++ src/include/api/flat_l2_index.h
@@ -161,7 +161,7 @@ class IndexFlatL2 {
 
     virtual void remove(const IdVector& ids) const = 0;
 
-    virtual size_t dimensions() const = 0;
+    virtual uint32_t dimensions() const = 0;
 
     virtual size_t num_vectors() const = 0;
   };
@@ -271,7 +271,7 @@ class IndexFlatL2 {
       //      index_.remove(ids);
     }
 
-    size_t dimensions() const override {
+    uint32_t dimensions() const override {
       return _cpo::dimensions(impl_index_);
     }
