+++ src/include/test/unit_ivf_pq_index.cc
@@ -581,7 +581,7 @@ TEST_CASE("query empty index", "[ivf_pq_index]") {
   tiledb::Context ctx;
   tiledb::VFS vfs(ctx);
   size_t num_vectors = 0;
-  size_t dimensions = 10;
+  uint32_t dimensions = 10;
   size_t nlist = 1;
   auto index = ivf_pq_index<siftsmall_feature_type, siftsmall_ids_type>(
       nlist, dimensions / 2);
@@ -639,7 +639,7 @@ TEST_CASE("query simple", "[ivf_pq_index]") {
   tiledb::VFS vfs(ctx);
 
   size_t num_vectors = 4;
-  size_t dimensions = 4;
+  uint32_t dimensions = 4;
   size_t nlist = 1;
   size_t num_subspaces = 2;
   size_t max_iter = 1;
@@ -712,7 +712,7 @@ TEST_CASE("ivf_pq_index query index written twice", "[ivf_pq_index]") {
   auto feature_type = "uint8";
   auto id_type = "uint32";
   auto partitioning_index_type = "uint32";
-  size_t dimensions = 3;
+  uint32_t dimensions = 3;
   size_t n_list = 1;
   size_t num_subspaces = 1;
   float convergence_tolerance = 0.00003f;
